+++ src/views/ApodView.vue
@@ -1,16 +1,23 @@
 <template>
   <div class="apod-view">
-    <ApodContainer />
+    <ErrorBoundary
+      componentName="ApodContainer"
+      errorMessage="Unable to load astronomy picture of the day page"
+    >
+      <ApodContainer />
+    </ErrorBoundary>
   </div>
 </template>
 
 <script>
+import ErrorBoundary from '@/components/common/ErrorBoundary.vue';
 import ApodContainer from '@/components/apod/ApodContainer.vue';
 
 export default {
   name: 'ApodView',
   title: 'Astronomy Picture of the Day',
   components: {
+    ErrorBoundary,
     ApodContainer,
   },
 };